Rewrite observation grouping (now sucks less)
