+++ docs/data/obs_select.rst
